+++ pkg/test/test_util.go
@@ -51,7 +51,7 @@ func NewPulsarClient() pulsar.Client {
 	return client
 }
 
-func AcquireUnusedPort() (int, error) {
+func AcquireUnusedPort() (uint16, error) {
 	tcpAddr, err := net.ResolveTCPAddr("tcp", "0.0.0.0:0")
 	if err != nil {
 		return 0, err
@@ -61,7 +61,7 @@ func AcquireUnusedPort() (int, error) {
 		return 0, err
 	}
 	defer l.Close()
-	return l.Addr().(*net.TCPAddr).Port, nil
+	return uint16(l.Addr().(*net.TCPAddr).Port), nil
 }
 
 func HttpGetRequest(url string) ([]byte, error) {
